[Fix] 에러 수정
 - Add gitignore file.
 - Rename library module.
 - Add missing .so files.
 - Update bluetooth permission for Android S or later.
 - Remove unnecessary dependencies.
 - Remove <use-feature> element from AndroidManifest.xml
