--- app/build.gradle
@@ -1,46 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-}
-
-android {
-    compileSdk 30
-
-    defaultConfig {
-        configurations.all { resolutionStrategy { force 'androidx.core:core-ktx:1.6.0' } }
-        minSdk 21
-        targetSdk 30
-        versionCode 1
-        versionName "1.0"
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-    kotlinOptions {
-        jvmTarget = '1.8'
-    }
-}
-
-dependencies {
-    implementation 'androidx.core:core-ktx:1.7.0'
-    implementation 'androidx.appcompat:appcompat:1.3.1'
-    implementation 'com.google.android.material:material:1.4.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.1.1'
-    implementation 'androidx.security:security-crypto:1.1.0-alpha03'
-    implementation 'gun0912.ted:tedpermission:2.2.3'
-    implementation 'com.google.code.gson:gson:2.8.9'
-    implementation files('libs/AicareUtils.jar')
-    implementation files('libs/AlgorithmUtil.jar')
-    implementation files('libs/getmorefatdata.jar')
-    implementation files('libs/iWeightLibrary.jar')
-}
