--- app/src/main/java/net/huray/solfit/bluetooth/data/enums/ScanState.kt
@@ -1,7 +1,0 @@
-package net.huray.solfit.bluetooth.data.enums
-
-enum class ScanState(value: Int) {
-    FAIL(0),
-    SCANNING(1),
-    STOPPED(2),
-}
