--- app/src/main/res/values/strings.xml
@@ -1,6 +1,0 @@
-<resources>
-    <string name="app_name">solfit_bluetooth_android</string>
-    <string name="error_denied_permission">권한을 거부하면 블루투스 서비스를 사용할 수 없습니다.\n\n[설정] > [권한]에서 권한을 설정하세요.</string>
-    <string name="error_bluetooth_not_enabled">블루투스 기능이 활성화되어있지 않습니다</string>
-    <string name="error_feature_not_supported">지원되지 않는 기능입니다</string>
-</resources>
