+++ libsolfit/.gitignore
