+++ libsolfit/proguard-rules.pro
