+++ libsolfit/src/main/java/net/huray/solfit/bluetooth/SolfitDataManager.kt
