+++ libsolfit/src/main/java/net/huray/solfit/bluetooth/data/MoreFatDataWrapper.kt
