+++ libsolfit/src/main/java/net/huray/solfit/bluetooth/data/UserInfo.kt
