+++ libsolfit/src/main/java/net/huray/solfit/bluetooth/data/enums/ScanState.kt
@@ -1,0 +1,7 @@
+package net.huray.solfit.bluetooth.data.enums
+
+enum class ScanState(val value: Int) {
+    FAIL(0),
+    SCANNING(1),
+    STOPPED(2),
+}
