+++ settings.gradle
@@ -1,2 +1,2 @@
 rootProject.name = "solfit_bluetooth_android"
-include ':app'
+include ':libsolfit'
